refactor(queue): extract exact review decision policy
